+++ src/test/java/kitchenpos/ui/ControllerAcceptanceTestHelper.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.product;
+package kitchenpos.ui;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
