+++ src/test/java/kitchenpos/ui/DatabaseCleaner.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.product;
+package kitchenpos.ui;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
